backend/requirements.txt
pillow
torch==2.3.1
torchvision==0.18.1
flask==3.0.3
flask-sqlalchemy
flask-bcrypt
python-dotenv
requests
flask-session
redis
uuid
datetime